+++ src/main/java/com/mobiquityinc/packer/model/Thing.java
@@ -1,12 +1,14 @@
 package com.mobiquityinc.packer.model;
 
+import java.math.BigDecimal;
+
 public class Thing {
 	
 	private Integer index;
-	private Double weight; 
-	private Double cost;
+	private BigDecimal weight; 
+	private BigDecimal cost;
 	
-	public Thing(Integer index, Double weight, Double cost) {
+	public Thing(Integer index, BigDecimal weight, BigDecimal cost) {
 		this.index = index;
 		this.weight = weight;
 		this.cost = cost;
@@ -20,19 +22,19 @@ public class Thing {
 		this.index = index;
 	}
 
-	public Double getWeight() {
+	public BigDecimal getWeight() {
 		return weight;
 	}
 
-	public void setWeight(Double weight) {
+	public void setWeight(BigDecimal weight) {
 		this.weight = weight;
 	}
 
-	public Double getCost() {
+	public BigDecimal getCost() {
 		return cost;
 	}
 
-	public void setCost(Double cost) {
+	public void setCost(BigDecimal cost) {
 		this.cost = cost;
 	}
